Apr 30 2019 by EqMule TBL placeholder info will again display correctly on the targetwindow if you use mq2targetinfo. Changed a bunch of stuff, plugin authors you have until next patch to update your plugins. If you ned help mail the plugin to me and ill send it back fixed: eqmule@hotmail.com   ALL global plugins in builder have been updated by me already, look at them for examples.   No more direct access to the window struct members in prep for moving to classes.   From now on call functions to get/put data out/in:   Example: too many to list see CSW struct.   I'll list 4 examples:  previously pWnd->XMLIndex now: pWnd->GetXMLIndex()                          previously pWnd->IsVisible()==false now: pWnd->IsVisible()==false        previously pWnd->dShow = true now: pWnd->SetVisible(true)        previously SetCXStr(&pWnd->WindowText,"blah"); now pWnd->CSetWindowText("blah");        previously GetCXStr(pWnd->WindowText,szOut); now GetCXStr(pWnd->CGetWindowText(),szOut); Added .NoExpendReagentID to the Spell tlo.   Usage: /echo ${Spell[Burst of Fire].NoExpendReagentID[x]} where x can be 1-4   This returns the ID of the needed Reagent you have to have in your inventory but will not be spent.

+++ EQBCS/ringbuffer.cpp
@@ -179,7 +179,7 @@ int RingBuffer::Read(unsigned char &in)
 	if (r_avail < 0)
 		r_avail = 0;
 	w_avail++;
-	if (w_avail > sz)
+	if (w_avail > (int)sz)
 		w_avail = sz;
 	full = false;
 	return 1;
